docker/bot/requirements.txt
python-telegram-bot==20.7
requests
cryptography
qdrant-client>=1.7.0
